feat: Default to a featured visualization (not Random ModFill)
