+++ src/shared/browserCaching.ts
@@ -1,3 +1,7 @@
+import {getFeatured} from './defineFeatured'
+import {math} from './math'
+import {specimenQuery} from './specimenEncoding'
+
 /* This file is responsible for all of the state of Numberscope that is kept
    in browser localStorage. Currently that consists of a collection of saved
    Specimens, and the list of IDs of OEIS sequences that will be shown in
@@ -35,122 +39,12 @@ function getCurrentDate(): string {
     return new Intl.DateTimeFormat('en-US', options).format(currentDate)
 }
 
-// QUERY ENCODING OF SPECIMENS
-const vizKey = 'viz'
-export const seqKey = 'seq'
-type QuerySpec = {
-    name: string
-    visualizerKind: string
-    sequenceKind: string
-    visualizerQuery: string
-    sequenceQuery: string
-}
-/**
- * Generates a URL query string from the information specifying a specimen.
- *
- * @param {string | QuerySpec} nameOrSpec
- *     The name of the specimen, or an object with key `name` and all of
- *     the other argument names as keys, in which case the other arguments
- *     are taken from this object instead
- * @param {string} visualizerKind  The kind of Visualizer
- * @param {string} sequenceKind  The kind of Sequence
- * @param {string?} visualizerQuery  Optional visualizer query parameter string
- * @param {string?} sequenceQuery  Optional sequence query parameter string
- * @return {string} the URL query string encoding of the parameter
- */
-export function specimenQuery(
-    nameOrSpec: string | QuerySpec,
-    visualizerKind?: string,
-    sequenceKind?: string,
-    visualizerQuery?: string,
-    sequenceQuery?: string
-): string {
-    let name = ''
-    if (!visualizerKind) {
-        // Only one arg, must be query
-        const spec = nameOrSpec as QuerySpec
-        name = spec.name
-        visualizerKind = spec.visualizerKind
-        sequenceKind = spec.sequenceKind
-        visualizerQuery = spec.visualizerQuery
-        sequenceQuery = spec.sequenceQuery
-    } else {
-        name = nameOrSpec as string
-    }
-    if (!sequenceKind) return ''
-    const leadQuery = new URLSearchParams({
-        name,
-        [vizKey]: visualizerKind,
-    })
-    const sepQuery = new URLSearchParams({[seqKey]: sequenceKind})
-    const queries = [leadQuery.toString()]
-    if (visualizerQuery) queries.push(visualizerQuery)
-    queries.push(sepQuery.toString())
-    if (sequenceQuery) queries.push(sequenceQuery)
-    return queries.join('&')
-}
-/**
- * Splits a URL query string for a specimen into its constituent parts
- * Returns an object with keys `name`, `visualizerKind`, `specimenKind`,
- * `visualizerQuery`, and `sequenceQuery`, corresponding to the five
- * arguments of specimenQuery(). I.e., this function inverts specimenQuery().
- *
- * @param {string} query  A URL query string encoding a specimen
- * @return {object} representation of components as decribed above.
- */
-export function parseSpecimenQuery(query: string) {
-    const params = new URLSearchParams(query)
-    const name = params.get('name') || 'Error: Unknown Name'
-    // We never insert a frame count in queries we generate, but
-    // we do parse it out in case it was specified, e.g. to make
-    // tests reproducible
-    const frames = parseFloat(params.get('frames') || 'Infinity')
-    // Similarly, we never insert a seed, but parse it in case it
-    // was specified
-    const seed = params.get('randomSeed') || null
-    const visualizerKind =
-        params.get(vizKey) || 'Error: No visualizer kind specified'
-    const sequenceKind =
-        params.get(seqKey) || 'Error: No sequence kind specified'
-    const vizPat = new RegExp(`&${vizKey}=[^&]*&`, 'd')
-    const seqPat = new RegExp(`&${seqKey}=[^&]*&?`, 'd')
-    let visualizerQuery = ''
-    let sequenceQuery = ''
-    const vizMatch = query.match(vizPat)
-    const seqMatch = query.match(seqPat)
-    if (vizMatch?.indices && seqMatch?.index && seqMatch?.indices) {
-        const firstAfterViz = vizMatch.indices[0][1]
-        if (seqMatch.index > firstAfterViz)
-            visualizerQuery = query.substring(firstAfterViz, seqMatch.index)
-        const firstAfterSeq = seqMatch.indices[0][1]
-        if (firstAfterSeq < query.length)
-            sequenceQuery = query.substring(firstAfterSeq)
-    }
-    return {
-        name,
-        frames,
-        seed,
-        visualizerKind,
-        sequenceKind,
-        visualizerQuery,
-        sequenceQuery,
-    }
-}
-
 // MEMORY RELATED HELPER FUNCTIONS AND VARIABLES
 
 // Keys of where the SIMs and IDs are saved (are arbitrary)
 const cacheKey = 'savedSpecimens'
 const currentKey = 'currentSpecimen'
 const idKey = 'activeIDs'
-
-// The default specimen
-// Will be displayed when the user visits the website for the first time
-export const defaultQuery = specimenQuery(
-    'Default Specimen',
-    'ModFill',
-    'Random'
-)
 
 // For backward compatibility:
 function newSIMfromOld(oldSim: {date: string; en64: string}): SIM {
@@ -252,7 +146,7 @@ export function getCurrent(): SIM {
         if ('en64' in data) return newSIMfromOld(data)
     }
 
-    return {query: defaultQuery, date: '', canDelete: true}
+    return math.pickRandom(getFeatured())
 }
 
 // Helper type for updateCurrent
